+++ test/test_base_s2v.py
@@ -24,10 +24,12 @@ from wordfreq import get_frequency_dict
 
 logger = logging.getLogger(__name__)
 
-CORPUS = Path("fse/test/test_data/test_sentences.txt")
+TEST_DATA = Path(__file__).parent / "test_data"
+CORPUS = TEST_DATA / "test_sentences.txt"
 DIM = 5
 W2V = Word2Vec(min_count=1, size=DIM)
-SENTENCES = [l.split() for i, l in enumerate(open(CORPUS, "r"))]
+with open(CORPUS, "r") as file:
+    SENTENCES = [l.split() for _, l in enumerate(file)]
 W2V.build_vocab(SENTENCES)
 
 
@@ -89,7 +91,7 @@ class TestBaseSentence2VecModelFunctions(unittest.TestCase):
 
     def test_save_load(self):
         se = BaseSentence2VecModel(W2V)
-        p = Path("fse/test/test_data/test_emb.model")
+        p = TEST_DATA / "test_emb.model"
         se.save(str(p.absolute()))
         self.assertTrue(p.exists())
         se2 = BaseSentence2VecModel.load(str(p.absolute()))
@@ -104,12 +106,12 @@ class TestBaseSentence2VecModelFunctions(unittest.TestCase):
         shape = (1000, 1000)
         ft.wv.vectors = np.zeros(shape, np.float32)
 
-        p = Path("fse/test/test_data/test_emb")
-        p_vecs = Path("fse/test/test_data/test_emb_wv.vectors")
-        p_ngrams = Path("fse/test/test_data/test_emb_ngrams.vectors")
-        p_vocab = Path("fse/test/test_data/test_emb_vocab.vectors")
+        p = TEST_DATA / "test_emb"
+        p_vecs = TEST_DATA / "test_emb_wv.vectors"
+        p_ngrams = TEST_DATA / "test_emb_ngrams.vectors"
+        p_vocab = TEST_DATA / "test_emb_vocab.vectors"
 
-        p_not_exists = Path("fse/test/test_data/test_emb.wv.vectors.npy")
+        p_not_exists = TEST_DATA / "test_emb.wv.vectors.npy"
 
         se = BaseSentence2VecModel(ft, wv_mapfile_path=str(p))
         self.assertTrue(p_vecs.exists())
@@ -132,10 +134,10 @@ class TestBaseSentence2VecModelFunctions(unittest.TestCase):
         ft = FastText(min_count=1, size=5)
         ft.build_vocab(SENTENCES)
 
-        p = Path("fse/test/test_data/test_emb")
-        p_vecs = Path("fse/test/test_data/test_emb_wv.vectors")
-        p_ngrams = Path("fse/test/test_data/test_emb_ngrams.vectors")
-        p_vocab = Path("fse/test/test_data/test_emb_vocab.vectors")
+        p = TEST_DATA / "test_emb"
+        p_vecs = TEST_DATA / "test_emb_wv.vectors"
+        p_ngrams = TEST_DATA / "test_emb_ngrams.vectors"
+        p_vocab = TEST_DATA / "test_emb_vocab.vectors"
 
         se = BaseSentence2VecModel(ft, wv_mapfile_path=str(p))
 
@@ -367,8 +369,8 @@ class TestBaseSentence2VecModelFunctions(unittest.TestCase):
 
     def test_move_ndarray_to_disk_w2v(self):
         se = BaseSentence2VecModel(W2V)
-        p = Path("fse/test/test_data/test_vecs")
-        p_target = Path("fse/test/test_data/test_vecs_wv.vectors")
+        p = TEST_DATA / "test_vecs"
+        p_target = TEST_DATA / "test_vecs_wv.vectors"
         se.wv.vectors[0, 1] = 10
         vecs = se.wv.vectors.copy()
         output = se._move_ndarray_to_disk(
@@ -385,9 +387,9 @@ class TestBaseSentence2VecModelFunctions(unittest.TestCase):
             output = se._move_ndarray_to_disk(se.wv.vectors)
 
     def test_move_w2v_vectors_to_disk_from_init(self):
-        p = Path("fse/test/test_data/test_vecs")
+        p = TEST_DATA / "test_vecs"
         se = BaseSentence2VecModel(W2V, wv_mapfile_path=str(p.absolute()))
-        p_target = Path("fse/test/test_data/test_vecs_wv.vectors")
+        p_target = TEST_DATA / "test_vecs_wv.vectors"
         self.assertTrue(p_target.exists())
         self.assertFalse(se.wv.vectors.flags.writeable)
         p_target.unlink()
@@ -396,10 +398,10 @@ class TestBaseSentence2VecModelFunctions(unittest.TestCase):
         ft = FastText(min_count=1, size=DIM)
         ft.build_vocab(SENTENCES)
 
-        p = Path("fse/test/test_data/test_vecs")
-        p_target_wv = Path("fse/test/test_data/test_vecs_wv.vectors")
-        p_target_ngram = Path("fse/test/test_data/test_vecs_ngrams.vectors")
-        p_target_vocab = Path("fse/test/test_data/test_vecs_vocab.vectors")
+        p = TEST_DATA / "test_vecs"
+        p_target_wv = TEST_DATA / "test_vecs_wv.vectors"
+        p_target_ngram = TEST_DATA / "test_vecs_ngrams.vectors"
+        p_target_vocab = TEST_DATA / "test_vecs_vocab.vectors"
 
         se = BaseSentence2VecModel(ft, wv_mapfile_path=str(p.absolute()))
 
@@ -521,8 +523,8 @@ class TestBaseSentence2VecPreparerFunctions(unittest.TestCase):
         self.assertTrue(se.sv.vectors_norm is None)
 
     def test_reset_vectors_memmap(self):
-        p = Path("fse/test/test_data/test_vectors")
-        p_target = Path("fse/test/test_data/test_vectors.vectors")
+        p = TEST_DATA / "test_vectors"
+        p_target = TEST_DATA / "test_vectors.vectors"
         se = BaseSentence2VecModel(W2V, sv_mapfile_path=str(p.absolute()))
         trainables = BaseSentence2VecPreparer()
         trainables.reset_vectors(se.sv, 20)
@@ -546,8 +548,8 @@ class TestBaseSentence2VecPreparerFunctions(unittest.TestCase):
         self.assertTrue(se.sv.vectors_norm is None)
 
     def test_update_vectors_memmap(self):
-        p = Path("fse/test/test_data/test_vectors")
-        p_target = Path("fse/test/test_data/test_vectors.vectors")
+        p = TEST_DATA / "test_vectors"
+        p_target = TEST_DATA / "test_vectors.vectors"
         se = BaseSentence2VecModel(W2V, sv_mapfile_path=str(p.absolute()))
         trainables = BaseSentence2VecPreparer()
         trainables.reset_vectors(se.sv, 20)
